fix: let the individual minecraft instance manage their own punishments
Old behaviour was to let the messages go through and then catch them. This behaviour is "too late" since messages are already through before the punishment is re-applied.
New behaviour is to let the instance themselves decide whether to re-apply the punishments to stop any message leakage from punished users.