fix(minecraft): fix deadlock when error during proxy connecting
This is a workaround to problems with proxy.
When proxy encounters a problem DURING the connecting phase, the instance will just enter a deadlock.
The only resolution is to pass an error and detect that specific error from the error handler side.

The error message is detected and handled at: ../handlers/error-handler.ts