refactor(core): set client-instance and event-handler accessors to protected
This is to enforce "Law of Demeter" by limiting the scope of all object properties as much as possible.
That way no outside intervention can change it in unpredictable manner.

All places that require the usage of these variables will receive it as an argument directly instead of having access to the parent/root object.