refactor: Introduce Instance and connectable-instance
Many classes and code required the same exact things: logger, error handler, application variable, etc.
"ClientInstance" used to refer to a bridge instance that connects and process information. But for utilities and smaller stuff, there was no proper way to unify all the definitions.

This patch introduces "Instance" which means anything that is stateful. "ConnectableInstance" is now the old "ClientInstance".