feat(minecraft): introduce dedicated send queue
With eventId present. old send function was getting too complicated and filled with bugs.
One of the bugs where ratelimiters stacking on top of each other or being executed multiple times on same command. the queue was also not properly followed due to multiple async awaits.

This solution will fix most of them and allow for further dedicated improvement.

How queue works:
- general commands are executed first
- guild action related commands are followed after
- chat messages are last.

This will increase the latency on chat but will allow for commands to load and not time out