fix(minecraft): change minecraft protocol version from 1.17.1 to 1.8.9
Hypixel has stopped supporting this version. The only two options were to either use the latest version or use 1.8.x version.

The latest version requires the latest nodejs version which not every user has or can install. So the old version is used.

There is no feature needed on the latest version as well. So, it won't be a compromise.