feat(core)!: make commands and minecraft chat events requiring originEventId
These events are only fired as a reaction to other events. It makes sense to set them that way.

This will also unify the "reply" method. It used to be that SOME events use the new originEventId system while other events such as commands use an old system (where the reply data is contained within the event itself).