refactor(core): create central migration file for the SQL system
This is useful later when trying to drop tables in some versions and maybe reusing the name later. Current system just assumes that no name is reused in any way.

This will also change the syntax from "IF NOT EXISTS" and ensures that the entire schema is fully ready by the time it is passed down to the individual components.